% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{This function helps us obtain reasonable global parameter to generate data to
do simulation}
\usage{
get_parameters(sample_weight)
}
\arguments{
\item{sample_weight}{the scale we multiply with Sigma_X, reflecting the observed sample
size of summary statistics of X ~ G}
}
\value{
A list of all the parameters
}
\description{
This function helps us obtain reasonable global parameter to generate data to
do simulation
}
