% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRR.R
\name{RRR}
\alias{RRR}
\title{This is the function to get naive reduced rank regression estimator for
the causal effect (Under random effect setting in MR)}
\usage{
RRR(Y, X, r, sqrt_Gamma = NULL, sqrt_Gamma_inv = NULL)
}
\arguments{
\item{Y}{Y in Y = CX + error}

\item{X}{X in Y = CX + error}

\item{r}{the specified reduced rank r}

\item{sqrt_Gamma}{Any weight matrix}

\item{sqrt_Gamma_inv}{Inverse of weight matrix}
}
\value{
A list of the estimated A, B and AB
}
\description{
This is the function to get naive reduced rank regression estimator for
the causal effect (Under random effect setting in MR)
}
