% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Generate data and run simulation}
\usage{
simulation(parameters, n)
}
\arguments{
\item{parameters}{get from get_parameters function, or we can also input anything we want}

\item{n}{number of instruments, p_z}
}
\value{
A list of calculated A_hat, B_hat, AB_hat, A_d_hat, B_d_hat, AB_d_hat
}
\description{
Generate data and run simulation
}
