% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt_matrix.R
\name{sqrt_matrix}
\alias{sqrt_matrix}
\title{Get the square root of a given matrix}
\usage{
sqrt_matrix(mat, inv = FALSE)
}
\arguments{
\item{mat}{the input matrix}

\item{inv}{do we want to return the inverse of square root}
}
\value{
(inverse of the) square root of the given matrix
}
\description{
Get the square root of a given matrix
}
